requirements.txt
flask
pymongo
python-dotenv
flask-jwt-extended
werkzeug
gunicorn
openai-agents
flask-cors